add timer, counter, spawner, group+member components add onCollide function to physics component refactor globals file to make adding new components easier, added macros to quickly access dense array addresses, component sizes, no. of components and max no. of components separate constants into constants.h add star item and starline struct, can spawn things in certain positions, from an array of mini-positions fix allocating a new entity id

+++ source/mem.c
@@ -36,7 +36,7 @@ int allocateObj(const u16* tiles, int tilesLen) {
             gObjAllocArr[i] = OBJ_SLOT_USED;
             for (i++; i < numBlocksReqd; i++)
                 gObjAllocArr[i] = OBJ_SLOT_CONTINUE;
-            memcpy32(MEM_VRAM_OBJ + (void*)(index * 32), tiles, tilesLen / 4);
+            memcpy32(MEM_VRAM_OBJ + (void*)(uint32_t)(index * 32), tiles, tilesLen / 4);
             return index;
         }
     }
